fix(spec-decode): package capability report command
